+++ src/harlequin_sqlite/adapter.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import sqlite3
+from contextlib import suppress
 from itertools import cycle, zip_longest
 from pathlib import Path
 from typing import Any, Literal, Sequence
@@ -60,15 +61,30 @@ class HarlequinSqliteConnection(HarlequinConnection):
     def __init__(self, conn: sqlite3.Connection, init_message: str = "") -> None:
         self.conn = conn
         self.init_message = init_message
-        self._auto_isolation_level = conn.isolation_level
-        self._transaction_mode_gen = cycle(["Auto", "Manual"])
+        self._transaction_modes = (
+            ["Auto", "Manual"] if hasattr(conn, "autocommit") else [""]
+        )
+        self._transaction_mode_gen = cycle(self._transaction_modes)
         self._transaction_mode = next(self._transaction_mode_gen)
         self._sync_connection_transaction_mode()
 
     def execute(self, query: str) -> HarlequinSqliteCursor | None:
+        # the behavior on manual mode is really counter-intuitive; if a
+        # transaction isn't explicitly began, it's basically the same as
+        # auto. By forcing an explicit begin, the behavior is more like
+        # manual mode on other databases.
+        if self.transaction_mode == "Manual" and not self.conn.in_transaction:
+            with suppress(sqlite3.Error):
+                self.conn.execute("begin;")
         try:
             cur = self.conn.execute(query)
         except sqlite3.Error as e:
+            # no-op if user ran a begin; statement, since we already started the txn
+            if (
+                isinstance(e, sqlite3.OperationalError)
+                and str(e) == "cannot start a transaction within a transaction"
+            ):
+                return None
             raise HarlequinQueryError(
                 msg=str(e),
                 title="SQLite raised an error when compiling or running your query:",
@@ -131,11 +147,20 @@ class HarlequinSqliteConnection(HarlequinConnection):
         self._sync_connection_transaction_mode()
         return new_mode
 
+    def close(self) -> None:
+        self.conn.close()
+
     def _sync_connection_transaction_mode(self) -> None:
+        if not self._transaction_mode or not hasattr(self.conn, "autocommit"):
+            return
+
+        if self.conn.in_transaction:
+            self.conn.commit()
+
         if self.transaction_mode == "Auto":
-            self.conn.isolation_level = self._auto_isolation_level
+            self.conn.autocommit = True
         else:
-            self.conn.isolation_level = None
+            self.conn.autocommit = False
 
     def _get_databases(self) -> list[str]:
         objects: list[tuple[str, str, str]] = self.conn.execute(
@@ -326,12 +351,10 @@ class HarlequinSqliteAdapter(HarlequinAdapter):
                 title="SQLite couldn't connect to your database.",
             ) from e
 
-        # Python 3.12 added an autocommit attribute that takes precendence over
-        # the isolation_level. We make sure to use the legacy behavior instead:
-        if hasattr(conn, "autocommit") and hasattr(
-            sqlite3, "LEGACY_TRANSACTION_CONTROL"
-        ):
-            conn.autocommit = sqlite3.LEGACY_TRANSACTION_CONTROL
+        # Python 3.12 added an autocommit attribute, which should be
+        # turned on by default
+        if hasattr(conn, "autocommit"):
+            conn.autocommit = True
 
         for uri, name in zip(other_dbs, db_names[1:]):
             try: